server/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
rich==13.5.2
starlette-exporter==0.19.0
prometheus_client==0.19.0
Cython==0.29.36
Pillow==10.3.0
pytest==8.3.2